+++ src/pages/NotFoundPage.tsx
@@ -1,0 +1,11 @@
+export const NotFoundPage = () => {
+  return (
+    <div>
+      <h1>404 Page Not Found</h1>
+      <div>Oops! The page you are looking for not found.</div>
+      <a href="/">
+        <span>Back to People List Page</span>
+      </a>
+    </div>
+  );
+};
